+++ http/user.go
@@ -5,10 +5,10 @@ import (
 )
 
 type UserStore interface {
-	Get(id int) (*teal.User, error)
+	Get(id int64) (*teal.User, error)
 	GetByUsername(username string) (*teal.User, error)
 	GetAll() ([]*teal.User, error)
 	Create(b *teal.User) (*teal.User, error)
-	Update(id int, b *teal.User) (*teal.User, error)
-	Delete(id int) error
+	Update(id int64, b *teal.User) (*teal.User, error)
+	Delete(id int64) error
 }
