+++ mock/mock.go
@@ -6,23 +6,23 @@ import (
 
 type BookStore struct {
 	GetAllBooksFn    func() ([]*teal.Book, error)
-	GetBookFn        func(id int) (*teal.Book, error)
+	GetBookFn        func(id int64) (*teal.Book, error)
 	GetBookByTitleFn func(title string) (*teal.Book, error)
 	CreateBookFn     func(b *teal.Book) (*teal.Book, error)
-	UpdateBookFn     func(id int, b *teal.Book) (*teal.Book, error)
-	DeleteBookFn     func(id int) error
+	UpdateBookFn     func(id int64, b *teal.Book) (*teal.Book, error)
+	DeleteBookFn     func(id int64) error
 	GetByAuthorFn    func(name string) ([]*teal.Book, error)
 }
 
 type AuthorStore struct {
-	GetAuthorFn     func(id int) (*teal.Author, error)
+	GetAuthorFn     func(id int64) (*teal.Author, error)
 	GetAllAuthorsFn func() ([]*teal.Author, error)
 	CreateAuthorFn  func(a *teal.Author) (*teal.Author, error)
-	UpdateAuthorFn  func(id int, a *teal.Author) (*teal.Author, error)
-	DeleteAuthorFn  func(id int) error
+	UpdateAuthorFn  func(id int64, a *teal.Author) (*teal.Author, error)
+	DeleteAuthorFn  func(id int64) error
 }
 
-func (m *BookStore) Get(id int) (*teal.Book, error) {
+func (m *BookStore) Get(id int64) (*teal.Book, error) {
 	return m.GetBookFn(id)
 }
 
@@ -38,11 +38,11 @@ func (m *BookStore) Create(b *teal.Book) (*teal.Book, error) {
 	return m.CreateBookFn(b)
 }
 
-func (m *BookStore) Update(id int, b *teal.Book) (*teal.Book, error) {
+func (m *BookStore) Update(id int64, b *teal.Book) (*teal.Book, error) {
 	return m.UpdateBookFn(id, b)
 }
 
-func (m *BookStore) Delete(id int) error {
+func (m *BookStore) Delete(id int64) error {
 	return m.DeleteBookFn(id)
 }
 
@@ -50,7 +50,7 @@ func (m *BookStore) GetByAuthor(name string) ([]*teal.Book, error) {
 	return m.GetByAuthorFn(name)
 }
 
-func (m *AuthorStore) Get(id int) (*teal.Author, error) {
+func (m *AuthorStore) Get(id int64) (*teal.Author, error) {
 	return m.GetAuthorFn(id)
 
 }
@@ -65,11 +65,11 @@ func (m *AuthorStore) Create(a *teal.Author) (*teal.Author, error) {
 
 }
 
-func (m *AuthorStore) Update(id int, a *teal.Author) (*teal.Author, error) {
+func (m *AuthorStore) Update(id int64, a *teal.Author) (*teal.Author, error) {
 	return m.UpdateAuthorFn(id, a)
 
 }
 
-func (m *AuthorStore) Delete(id int) error {
+func (m *AuthorStore) Delete(id int64) error {
 	return m.DeleteAuthorFn(id)
 }
